Move content of runner to the new count
